pyproject.toml
[tool.poetry]
name = "avto-discord-bot"
version = "0.1.0"
description = "A discord bot for notifying about new listings on the avto.net website."
authors = ["mevljas <sebastjan.mevlja@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.7"
beautifulsoup4 = "^4.12.3"
black = "24.10.0"
playwright = "^1.45.1"
sqlalchemy = "^2.0.31"
pytest-playwright = "^0.5.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
pylint = "^3.3.1"
urllib3 = "^2.2.3"
aiosqlite = "^0.20.0"
discord = "^2.3.2"
pre-commit = "^4.0.1"

[tool.black]
required-version = "24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

